# product-api/swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
    Product:
        description: Product Product defines the structure for an API product
        properties:
            description:
                description: description
                maxLength: 10000
                type: string
                x-go-name: Description
            id:
                description: ID
                format: int64
                minimum: 1
                type: integer
                x-go-name: ID
            name:
                description: name
                maxLength: 255
                type: string
                x-go-name: Name
            price:
                description: price
                format: float
                minimum: 0.01
                type: number
                x-go-name: Price
            sku:
                description: s k u
                pattern: '[a-z]+-[a-z]+-[a-z]+'
                type: string
                x-go-name: SKU
        required:
            - id
            - name
            - price
            - sku
        type: object
        x-go-package: github.com/hnsia/go-nic/product-api/client/models
info:
    description: Documentation for Product API
    title: of Product API
    version: 1.0.0
paths:
    /products:
        get:
            description: Returns a list of products
            operationId: listProducts
            parameters:
                - description: |-
                    Currency used when returning the price of the product,
                    when not specified, currency is returned in GBP.
                  in: query
                  name: Currency
                  type: string
            responses:
                "200":
                    $ref: '#/responses/productsResponse'
            tags:
                - products
    /products/{id}:
        delete:
            description: Returns nothing
            operationId: deleteProduct
            parameters:
                - description: The id of the product to delete from the database
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "201":
                    $ref: '#/responses/noContent'
            tags:
                - products
produces:
    - application/json
responses:
    noContent:
        description: ""
    productsResponse:
        description: A list of products returns in the response
        schema:
            items:
                $ref: '#/definitions/Product'
            type: array
schemes:
    - http
swagger: "2.0"